Changing variable type from int to double

+++ OpenWeatherAPITesting/Models/Wind.cs
@@ -12,6 +12,6 @@ namespace OpenWeatherAPITesting.Models
     public class Wind
     {
         public double speed;
-        public int deg;
+        public double deg;
     }
 }
